- fixed an import typo in example_use_of_drug_campaign.py
- fixed a typo in add_immune_overlays in demographics.py

- fixed other minor bugs in demographics.py to allow the addition of
immune overlays 

- RunSimulations of SimulationManager calls the set of functions in
mod_fn_list. If one of these functions is site_fn(v),
the value of paramater "Demographics_Filename" got overwritten by the
default value of "Demographics_Filename" in geography.py
for the given _site_ (prepended with the path to the corresponding
demographics file).

This behavior causes errors when "Demographics_Filename" contains a set
of files (e.g. immune overlays) prior to the overwrite.
The current fix for the behavior is a bit ugly but works, for now. If
"Demographics_Filename" already contains a ';' separated list of files
in geography.py/set_geography(...), these files are prepended with the
proper data input path, but are not overwritten.

In general the current behavior of mod_fn's called in the
SimulationManager may lead to similar issues. When a set of parameters
are updated in a setup script (such as example_sweep.py) by calling
cb.update(...), and some or all of these parameters are also updated via 
functions in mod_fn_list, the cb.update(...) values would be
overwritten.